Support HiddenComponent and LegacyHiddenComponent in test selectors
Typically things are hidden this way and then they wouldn't show up as
hidden through a prop.

However, if something is hidden with the DOM prop we can also support that
case. Things can be hidden in other ways too that makes them not focusable
etc, but to simplify we recommend the hidden prop.

I avoid passing the Fiber to the Host config so it remains opaque. See
shouldSetTextContent for precedence for this pattern for example.