Unify Server / Client Reference Encoding
Server References are currently outlined and then referenced by a special
reference instead of a special row. The idea was that it would mostly just
be a single ID string instead of a complex object so it would typically
be inlined and there's no reason to eagerly load one.

But it's simpler to unify them since they're currently always outlined
anyway. Partly due to the bound arguments and debug data.